+++ src/app/oic/pages/plt-codigo-etica/plt-codigo-etica.component.html
@@ -1,18 +1,21 @@
 <div class="cdo-etica">
     <div class="message-fider">
-      <app-finder-oic (onEnter)="loadTreefromFinder($event)"
-        (onSelectedOicChange)="loadTreeFromFinderSelectedOic($event)"></app-finder-oic>
+      <app-finder-oic (onEnter)="showCodigo(  $event  )"
+        (onSelectedOicChange)="showCodigo2(  $event  )"
+        [title]="'Buscador de entes publicos'"
+        ></app-finder-oic>
      
     </div>
   
-    <p-divider id="divider-user" layout="vertical" align="center">
+    <p-divider id="divider-user" layout="vertical" align="center" >
       <i class="pi pi-user mr-2"></i>
     </p-divider>
   
-  
-    <app-tree [data]="data" (onNodeSelected)="viewUserData( $event )"></app-tree>
+    <div *ngIf="display">
+      <button pButton type="button"class="p-button-outlined">
+        <a href="http://www.africau.edu/images/default/sample.pdf">Descargar Código de Ética de {{ name_ente }}</a>
+      </button>
+    </div>
   
   </div>
-  <p-dialog header="Datos de contacto" [(visible)]="display">
-    <app-view-users [users]="selectedUser"></app-view-users>
-  </p-dialog>
+
