+++ src/app/oic/pages/plt-codigo-etica/plt-codigo-etica.component.ts
@@ -1,8 +1,8 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { MessageService, TreeNode } from 'primeng/api';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { MessageService } from 'primeng/api';
 import { AuthService } from 'src/app/auth/services/auth.service';
-import { user_edit } from 'src/app/shared/models/colaborador.interface';
 import { OicEnte, OicInterface } from '../../models/oic.interface';
+
 import { filterBoss, filterEnte } from '../../models/tree.interface';
 
 @Component({
@@ -14,81 +14,31 @@ import { filterBoss, filterEnte } from '../../models/tree.interface';
 })
 export class PltCodigoEticaComponent  {
 
-  data: TreeNode[] = [];
-  filter: filterBoss | filterEnte | null = null;
  
-  ente:string = '';
-  haveError:boolean = false;
+  filter: filterBoss | filterEnte | null = null; 
   display: boolean = false;
-  selectedUser:user_edit = {} as user_edit;
+  name_ente: string = "";
+ 
 
   constructor(
     private readonly as: AuthService,
     private readonly ms: MessageService
   ) { } 
-
-  loadTreefromFinder( $event:OicEnte){
-    this.ente = $event.ente.id;
-    
-    const params: filterEnte = {
-      boss: {
-        ente: $event.ente.id
-      }
-    }
-    this.loadTree(params);
-    //console.log("this.ente desde hijo",$event.ente.id);  
-  }
-  
-  loadTreeFromFinderSelectedOic( $event:OicInterface){
-    this.ente = $event.id;
-    
-    const params: filterEnte = {
-      boss: {
-        ente: $event.id
-      }
-    }
-    this.loadTree(params);
-    //console.log("this.ente desde hijo",$event.id);
-  }
-
-  loadTree(params: filterBoss | filterEnte) {
-    //console.log(">>>>>>>>>>>CARGANDO INFORMACION DEL ARBOL>>>>>>>>>>>>>>>>>>>>", params);
-    this.as.get_tree_colaboradores(params).subscribe({
-      next: (result) => {
-        const tree: any = result.data!;
-        //console.log(">>>>>>>>>>>>>>>>>>>>>>>>>>>",tree);
-        if(tree.getColaboresTreeData.data !== null){
-          this.data = [tree.getColaboresTreeData];
-        }
-        else{
-          this.showError();
-          this.purgeTree();
-        }
-  
-      },
-      error: (err) => {
-        //console.log(">>>>>>>>>>>>>>>>>>>>>>>>>>>Error en la consulta",err);
-      }
-    });
-  
-  }
-  
-  purgeTree() {
-    this.data = [];
-  }
-  
-  viewUserData( user: user_edit ){
-    //console.log(">>>>>>>>>>>>>>>",user);
-    this.selectedUser = user;
-    this.showPopup();
-  }
   
   showError() {
     this.ms.add({ severity: 'info', summary: 'Información', detail: 'No hay datos del ente solicitado...' });   //<-- Mensaje de error
   }
-  showPopup(){
+  showCodigo($event: OicEnte) {
+    
+    console.log(">>>>>>>>>>>>>>>>>>>>>>>>>>>", $event);
     this.display = true;
+    this.name_ente = $event.ente.nombre_ente;
   }
-  
+
+  showCodigo2($event: OicInterface) {
+    console.log(">>>>>>>>>>>>>>>>>>>>>>>>>>>", $event);
+    this.display = true;
+    this.name_ente = $event.nombre_ente;
+  }
 
 }
