+++ src/app/oic/pages/plt-contacto/plt-contacto.component.html
@@ -1,1 +1,1 @@
-<p>plt-contacto works!</p>
+<app-view-conctacto></app-view-conctacto>
