+++ src/app/oic/pages/plt-home/plt-home.component.html
@@ -1,13 +1,16 @@
 <div class="home-plt container">
     <p-fieldset legend="Modelo para la Construcción de los Códigos de Ética" [toggleable]="true" >
        <p class="txt-inf">
-       Lorem ipsum dolor sit amet, consectetur adipisicing elit. 
-       Provident distinctio eveniet eius? Ad eaque saepe fugiat, enim nostrum libero
-       mollitia sapiente velit id eius harum quis magni corrupti nemo ipsa.
-       Lorem ipsum dolor, sit amet consectetur adipisicing elit.
-       Consequuntur earum eaque facilis aut quam nemo, quidem nisi 
-       accusantium voluptatem consequatur necessitatibus illum maxime 
-       quo laudantium praesentium minima iusto fugiat totam.
+        Ética pública
+        Disciplina basada en normas de
+        conducta que se fundamentan en
+        el deber público y que busca en
+        toda decisión y acción, la
+        prevalencia del bienestar de la
+        sociedad en coordinación con los
+        objetivos del Estado mexicano, de
+        los entes públicos y de la
+        responsabilidad de la persona ante éstos.
        </p>
        <button pButton type="button"class="p-button-outlined">
         <a href="http://www.africau.edu/images/default/sample.pdf">Descargar Modelo para la Construcción de los Códigos de Ética </a>
