+++ src/app/oic/pages/plt-preguntas/plt-preguntas.component.html
@@ -1,1 +1,22 @@
-<p>plt-preguntas works!</p>
+<h3>Preguntas frecuentes</h3>
+<br>
+<div class="home-plt container">
+    <p-fieldset legend="¿Qué entiendes por ética pública?" [toggleable]="true" [collapsed]="false"  >
+       <p class="txt-inf">         
+            “La ética pública son las relaciones que se producen entre personas, colectivos e instituciones, en la esfera pública, que se basan en criterios de justicia y de respeto por las normas convenidas socialmente”.
+       </p>
+    </p-fieldset>
+    <p-fieldset legend="¿Cuáles son los principios de la ética pública?" [toggleable]="true" [collapsed]="false" >
+        <p class="txt-inf">     
+            Principios de la Ética Pública
+
+            b) Probidad Actuar con integridad, rectitud y honradez. c) Igualdad Tratar a todas las personas por igual en condiciones similares. d) Imparcialidad Proceder con objetividad en el ejercicio de la función pública. e) Justicia Dar a cada quien lo que le corresponde, según derecho y razón.
+        </p>
+     </p-fieldset>
+     <p-fieldset legend="¿Cómo se aplica la ética pública?" [toggleable]="true" [collapsed]="false" >
+        <p class="txt-inf">
+            En la práctica, la ética en la función pública nos permite solucionar conflictos, crear puentes entre posturas contrarias en una sociedad de tantos contrastes como es en la que vivimos. La ética nos permite además innovar y crear soluciones cuando ni la tradición ni las autoridades son una opción.
+        </p>
+     </p-fieldset>
+     
+</div>
