+++ src/app/oic/pages/plt-preguntas/plt-preguntas.component.scss
@@ -1,0 +1,16 @@
+.home-plt{
+    padding-bottom: 10rem;
+}
+
+.txt-home-plt{
+    font-size: 1.5rem;
+    font-weight: bold;
+    color: #fff;
+    text-align: center;
+    margin-bottom: 2rem;
+}
+.txt-inf{
+    font-size: 1.4rem;
+        
+    margin-bottom: 2rem;
+}
