+++ src/app/shared/pages/view-conctacto/view-conctacto.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-view-conctacto',
+  templateUrl: './view-conctacto.component.html',
+  styleUrls: ['./view-conctacto.component.scss']
+})
+export class ViewConctactoComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
